SelfTest: Consolidate local and global selftest
This commit hoists out the common code between local and global
selftest into a helper class SelfTest.
